+++ contracts/stakeERC20.sol
@@ -29,7 +29,7 @@ contract StakeERC20 is StakeBase {
         uint24 weekSeconds,
         uint8 maxWeeks,
         uint256 maxScale,
-        uint8 lockWeeks,
+        uint16 lockWeeks,
         uint256 beginTime,
         uint256 endTime,
         address _rewardToken,
